requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
ffmpeg-python==0.2.0
